fix bug delete

+++ src/main/java/fr/unsplash/repository/ImageUnsplashService.java
@@ -28,7 +28,7 @@ public class ImageUnsplashService implements IImageUnsplashService{
 
     @Override
     public boolean deleteById(Long id) {
-        long nbLinesRemoved = dao.deleteImageById(id);
+        int nbLinesRemoved = dao.deleteImageById(id);
         return nbLinesRemoved != 0;
     }
